requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiokafka==0.12.0
aiosignal==1.4.0
altair==5.5.0
async-timeout==5.0.1
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.0
colorama==0.4.6
frozenlist==1.7.0
gevent==25.5.1
gitdb==4.0.12
GitPython==3.1.44
greenlet==3.2.3
idna==3.10
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
kafka-python==2.2.6
MarkupSafe==3.0.2
multidict==6.6.3
narwhals==1.38.2
newsapi-python==0.2.7
numpy==2.2.5
packaging==24.2
pandas==2.2.3
pillow==11.2.1
propcache==0.3.2
protobuf==6.30.2
psycopg2-binary==2.9.10
py4j==0.10.9.9
pyarrow==20.0.0
pycparser==2.22
pydeck==0.9.1
pyspark==4.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
rpds-py==0.24.0
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
setuptools==80.9.0
six==1.17.0
smmap==5.0.2
streamlit==1.45.0
tenacity==9.1.2
toml==0.10.2
tornado==6.4.2
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
watchdog==6.0.0
websocket==0.2.1
websockets==15.0.1
yarl==1.20.1
zope.event==5.1.1
zope.interface==7.2
